relocation components srp
