+++ src/client/src/containers/PageAzureLogin/index.tsx
@@ -17,7 +17,7 @@ import { setDetailPageAction } from "../../actions/wizardInfoActions/setDetailsP
 import { IOption } from "../../types/option";
 import { azureMessages } from "../../mockData/azureServiceOptions";
 import { AppState } from "../../reducers";
-import AzureSubscriptions from "../AzureSubscriptions";
+import AzureSubscriptions from "./AzureSubscriptions";
 import AzureStudent from "./AzureStudent";
 import Title from "../../components/Title";
 import RootAction from "../../actions/ActionType";
