+++ src/client/src/containers/RightSidebar/AppServiceSelection/index.tsx
@@ -1,23 +1,23 @@
 import * as React from "react";
 import { connect } from "react-redux";
 
-import * as getSvg from "../../utils/getSvgUrl";
+import * as getSvg from "../../../utils/getSvgUrl";
 
-import DraggableSidebarItem from "../../components/DraggableSidebarItem";
+import DraggableSidebarItem from "../../../components/DraggableSidebarItem";
 
-import { removeAppServiceSettingsAction } from "../../actions/azureActions/appServiceActions";
-import { IAppService } from "../../reducers/wizardSelectionReducers/services/appServiceReducer";
-import { ReactComponent as EditIcon } from "../../assets/edit.svg";
+import { removeAppServiceSettingsAction } from "../../../actions/azureActions/appServiceActions";
+import { IAppService } from "../../../reducers/wizardSelectionReducers/services/appServiceReducer";
+import { ReactComponent as EditIcon } from "../../../assets/edit.svg";
 
-import { openAppServiceModalAction } from "../../actions/modalActions/modalActions";
+import { openAppServiceModalAction } from "../../../actions/modalActions/modalActions";
 
 import styles from "./styles.module.css";
-import { KEY_EVENTS } from "../../utils/constants";
+import { KEY_EVENTS } from "../../../utils/constants";
 
 import { injectIntl, InjectedIntlProps } from "react-intl";
 import { ThunkDispatch } from "redux-thunk";
-import { AppState } from "../../reducers";
-import RootAction from "../../actions/ActionType";
+import { AppState } from "../../../reducers";
+import RootAction from "../../../actions/ActionType";
 
 interface IProps {
   appServiceSelection: IAppService;
